Move patchArrayBuffer to partialIO
