+++ packages/typegpu/src/indexNamedExports.ts
@@ -27,11 +27,8 @@ export { isTgpuComputeFn } from './core/function/tgpuComputeFn.ts';
 export { isVariable } from './core/variable/tgpuVariable.ts';
 export { ShaderGenerator } from './tgsl/shaderGenerator.ts';
 export { WgslGenerator } from './tgsl/wgslGenerator.ts';
-export {
-  patchArrayBuffer,
-  readFromArrayBuffer,
-  writeToArrayBuffer,
-} from './core/buffer/arrayBufferIO.ts';
+export { readFromArrayBuffer, writeToArrayBuffer } from './core/buffer/arrayBufferIO.ts';
+export { patchArrayBuffer } from './data/partialIO.ts';
 
 // types
